request

+++ Pages/Book/Books.cshtml.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using AutoMapper;
 using DBook.Data;
 using DBook.Models;
+using DBook.Services;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
@@ -16,17 +17,24 @@ namespace DBook.Pages.Book
     {
         private ApplicationDbContext _ctx;
         private IMapper _mapper;
+       
+        
+        public ILookup<string, Data.Entity.Book> books { set; get; }
 
         public BooksModel(ApplicationDbContext context,IMapper mapper)
         {
             _ctx = context;
             _mapper=mapper;
+           
         }
         public IEnumerable<BookModel> Books { get; set; }
         public async Task<IActionResult> OnGetAsync()
         {
               var result = await _ctx.Books.Select(x=>x).ToListAsync();
                Books = _mapper.Map<IEnumerable<Data.Entity.Book>, IEnumerable<BookModel>>(result);
+               books =   _ctx.Books.Select(e => e).ToLookup(e => e.PublisherId);
+               var userId = User.GetUserId();
+               ViewData["success"] = "no";
             return Page();
         }
 
@@ -34,12 +42,24 @@ namespace DBook.Pages.Book
         public IActionResult OnPost(int id)
         {
            var book = _ctx.Books.FirstOrDefault(x=>x.Id==id);
+           var userId = User.GetUserId();
           if (book == null)
                return NotFound();
-            _ctx.Books.Remove(book);
-            _ctx.SaveChanges();
+          if (userId != book.PublisherId)
+          {
+                ViewData["success"] = false;
+                return RedirectToPage("/Book/books");
+          }
 
+          _ctx.Books.Remove(book);
+            _ctx.SaveChanges();
+            ViewData["success"] = true;
             return RedirectToPage("/Book/books");
+        }
+
+        public async Task<IActionResult> OnRequestAsync(BookModel model)
+        {
+            return Page();
         }
 
     }
